+++ src/main/java/baseball/game/BaseballGame.java
@@ -8,7 +8,7 @@ import static baseball.constant.GameConstant.*;
 import static baseball.constant.GameMessage.*;
 import static baseball.game.CompareNumbers.*;
 import static baseball.number.ComputerNumber.createComputerNumber;
-import static baseball.exception.RestartNumberValidator.checkValidationRestartNumber;
+import static baseball.validator.RestartNumberValidator.checkValidationRestartNumber;
 import static baseball.number.UserNumber.checkInputUserNumber;
 
 public class BaseballGame {
