+++ src/test/java/baseball/validator/UserNumberValidatorTest.java
@@ -1,4 +1,4 @@
-package baseball.exception;
+package baseball.validator;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -7,8 +7,6 @@ import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class UserNumberValidatorTest {
